+++ src/android/com/outsystems/firebase/analytics/validator/OSFANLMinimumRequired.kt
@@ -1,0 +1,8 @@
+package com.outsystems.firebase.analytics.validator
+
+/**
+ * Represents the types of minimum required items validation
+ */
+enum class OSFANLMinimumRequired {
+    ONE, AT_LEAST_ONE, NONE
+}