+++ src/ios/Common/OSFANLDefaultValues.swift
@@ -1,0 +1,4 @@
+struct OSFANLDefaultValues {
+    static let itemCustomParametersMaximum: Int = 27
+    static let eventItemsMaximum: Int = 200
+}