+++ src/ios/FirebaseAnalyticsPlugin.h
@@ -3,6 +3,7 @@
 @interface FirebaseAnalyticsPlugin : CDVPlugin
 
 - (void)logEvent:(CDVInvokedUrlCommand*)command;
+- (void)logECommerceEvent:(CDVInvokedUrlCommand *)command;
 - (void)setUserId:(CDVInvokedUrlCommand*)command;
 - (void)setUserProperty:(CDVInvokedUrlCommand*)command;
 - (void)setEnabled:(CDVInvokedUrlCommand*)command;