+++ src/ios/InputTransformer/OSFANLInputTransformable.swift
@@ -1,0 +1,3 @@
+protocol OSFANLInputTransformable {
+    func transform(_ eventParameterArray: [InputParameterData]?, _ itemArray: [InputItemData]?) throws -> OSFANLInputTransformableModel
+}